+++ lib/auth.js
@@ -19,14 +19,10 @@ async function createUser(email, password) {
       uid: uid,
       email: email,
     })
-    sendEmailVerification(userCredential.user)
-    .then(() => {
-      // Email verification sent!
-      // ...
-    });
+    // ...
   })
   .catch((error) => {
-    const errorMessage = error.code;
+    const errorMessage = error.message;
     return handleErrors2(errorMessage);
     // ..
   });
@@ -45,7 +41,7 @@ async function signIn(email, password) {
   })
   .catch((error) => {
     // Handle Errors here.
-    const errorMessage = error.code;
+    const errorMessage = error.message;
     return handleErrors(errorMessage);
   });
 }
