+++ package.json
@@ -8,28 +8,24 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@emotion/react": "^11.8.2",
-    "@emotion/styled": "^11.8.1",
-    "@fontsource/roboto": "^4.5.3",
+    "@emotion/react": "^11.7.1",
+    "@emotion/styled": "^11.6.0",
     "@material-ui/core": "^4.12.3",
     "@material-ui/icons": "^4.11.2",
-    "@mui/icons-material": "^5.5.0",
-    "@mui/material": "^5.5.0",
-    "@mui/styled-engine-sc": "^5.4.2",
+    "@mui/icons-material": "^5.4.2",
+    "@mui/material": "^5.4.1",
     "@mui/styles": "^5.4.2",
     "axios": "^0.26.0",
     "firebase": "^9.6.3",
     "firebase-admin": "^10.0.2",
     "formik": "^2.2.9",
     "next": "^12.0.7",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "17.0.2",
+    "react-dom": "17.0.2",
     "react-gsap": "^3.2.1",
-    "react-password-strength-bar": "^0.4.0",
     "react-scrollmagic": "^2.3.0",
     "react-table": "^7.7.0",
     "sass": "^1.48.0",
-    "styled-components": "^5.3.3",
     "vercel": "^24.0.0"
   },
   "devDependencies": {
