+++ pages/Account/About.js
@@ -1,5 +1,6 @@
 import React from 'react'
-import styles from './NewAccount.module.scss';
+import styles from '../../styles/Home.module.scss'
+import ActMenu from './ActMenu';
 import Card from '@material-ui/core/Card';
 import CardActionArea from '@material-ui/core/CardActionArea';
 import CardActions from '@material-ui/core/CardActions';
@@ -11,10 +12,11 @@ import LoggedIn from '../LoggedIn';
 import logo from '../../public/1kartCircle.jpg';
 import Image from 'next/image'
 
-export default function Support() {
+export default function ActAbout() {
     return(
         <LoggedIn>
         <div className = {styles.container}>
+            <ActMenu/>
             <div className = {styles.about}>
                 <h2>About Us</h2>
                 {/*<Image src={logo} alt="logo" className={styles.image}/>*/}
