+++ pages/Account/History.js
@@ -1,13 +1,15 @@
 import React from 'react'
-import styles from './NewAccount.module.scss';
+import styles from '../../styles/Home.module.scss'
+import ActMenu from './ActMenu';
 import LoggedIn from '../LoggedIn';
 import logo from '../../public/1kartCircle.jpg';
 import Image from 'next/image'
 
-export default function OrderHistory() {
+export default function ActHist() {
     return(
         <LoggedIn>
         <div className = {styles.container}>
+            <ActMenu />
             <div className = {styles.history}>
                 <h2>Coming Soon</h2>
                 {/*<Image src={logo} width="70px" height="70px" alt="logo" className={styles.image}/>*/}
