--- pages/Account/NewAccount.module.scss
@@ -1,361 +1,0 @@
-@use 'variables';
-
-.container {
-    padding: 0 2rem;
-}
-
-.container h2 {
-    padding-top: 1rem;
-    margin-left: 2rem;
-}
-
-/* Index */
-.page {
-    padding-top: 6em;
-    font-family: "Montserrat", san-serif;
-    display: grid;
-    text-align: center;
-    row-gap: 2rem;
-    padding-bottom: 15rem;
-}
-
-.btnGrid1 {
-    grid-column-start: 1;
-    grid-column-end: 2;
-    grid-row-start: 1;
-    grid-row-end: 2;
-}
-
-.infoBtn {
-    background-color: rgb(182, 220, 231);
-    height: 6rem;
-    width: 18rem;
-    border: none;
-    border-radius: .5rem;
-    font-family: "Montserrat", san-serif;
-    font-size: 1.1rem;
-}
-
-.btnGrid2 {
-    grid-column-start: 2;
-    grid-column-end: 3;
-    grid-row-start: 1;
-    grid-row-end: 2;
-}
-
-.orderBtn {
-    background-color: rgb(182, 220, 231);
-    height: 6rem;
-    width: 18rem;
-    border: none;
-    border-radius: .5rem;
-    font-family: "Montserrat", san-serif;
-    font-size: 1.2rem;
-}
-
-.btnGrid3 {
-    grid-column-start: 3;
-    grid-column-end: 4;
-    grid-row-start: 1;
-    grid-row-end: 2;
-}
-
-.listsBtn {
-    background-color: rgb(182, 220, 231);
-    height: 6rem;
-    width: 18rem;
-    border: none;
-    border-radius: .5rem;
-    font-family: "Montserrat", san-serif;
-    font-size: 1.2rem;
-}
-
-.btnGrid4 {
-    grid-column-start: 1;
-    grid-column-end: 2;
-    grid-row-start: 2;
-    grid-row-end: 3;
-}
-
-.warrantyBtn {
-    background-color: rgb(182, 220, 231);
-    height: 6rem;
-    width: 18rem;
-    border: none;
-    border-radius: .5rem;
-    font-family: "Montserrat", san-serif;
-    font-size: 1.3rem;
-}
-
-.btnGrid5 {
-    grid-column-start: 2;
-    grid-column-end: 3;
-    grid-row-start: 2;
-    grid-row-end: 3;
-}
-
-.addressBtn {
-    background-color: rgb(182, 220, 231);
-    height: 6rem;
-    width: 18rem;
-    border: none;
-    border-radius: .5rem;
-    font-family: "Montserrat", san-serif;
-    font-size: 1.1rem;
-}
-
-.btnGrid6 {
-    grid-column-start: 3;
-    grid-column-end: 4;
-    grid-row-start: 2;
-    grid-row-end: 3;
-}
-
-.supportBtn {
-    background-color: rgb(182, 220, 231);
-    height: 6rem;
-    width: 18rem;
-    border: none;
-    border-radius: .5rem;
-    font-family: "Montserrat", san-serif;
-    font-size: 1.3rem;
-}
-
-/* Profile (Account Information) */
-  
-.menu1 {
-    height: 100vh;
-    text-align: center;
-    width: 15%;
-    float: left;
-    background: linear-gradient(rgb(244, 250, 255), rgb(255, 255, 255));
-}
-  
-.tabs {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-    height: 10%;
-}
-  
-.profile {
-    float: left;
-    height: 100vh;
-    width: 80%;
-    display: grid;
-    gap: 1em 1em;
-    padding-left: 3em;
-}
-  
-.grid1 {
-    grid-column-start: 1;
-    grid-column-end: 2;
-    grid-row-start: 1;
-    grid-row-end: 2;
-    padding-top: 3em;
-}
-  
-.grid2 {
-    grid-column-start: 1;
-    grid-column-end: 2;
-    grid-row-start: 2;
-    grid-row-end: 3;
-}
-  
-.grid3 {
-    grid-column-start: 2;
-    grid-column-end: 3;
-    grid-row-start: 2;
-    grid-row-end: 3;
-}
-  
-.grid4 {
-    grid-column-start: 3;
-    grid-column-end: 4;
-    grid-row-start: 2;
-    grid-row-end: 3;
-}
-  
-.grid5 {
-    grid-column-start: 1;
-    grid-column-end: 2;
-    grid-row-start: 3;
-    grid-row-end: 4;
-}
-  
-.grid6 {
-    grid-column-start: 2;
-    grid-column-end: 3;
-    grid-row-start: 3;
-    grid-row-end: 4;
-}
-  
-.grid7 {
-    grid-column-start: 3;
-    grid-column-end: 4;
-    grid-row-start: 3;
-    grid-row-end: 4;
-}
-  
-.grid8 {
-    grid-column-start: 1;
-    grid-column-end: 2;
-    grid-row-start: 4;
-    grid-row-end: 5;
-    align-content: center;
-}
-  
-.grid9 {
-    grid-column-start: 3;
-    grid-column-end: 4;
-    grid-row-start: 1;
-    grid-row-end: 2;
-    padding-top: 1em;
-}
-  
-.kartLogo {
-    width: 150px;
-    height: 150px;
-}
-  
-.grid1 h2 {
-    font-family: 'Roboto', sans-serif;
-    font-size: 2rem;
-    margin-left: -.5rem;
-}
-  
-.grid2, .grid3, .grid4, .grid5, .grid6, .grid7, .grid8 label {
-    font-family: 'Roboto', sans-serif;
-}
-  
-.updateBtn:hover {
-    background-color: white;
-    color: blue;
-}
-
-/* Order History */
-
-.history {
-    float: left;
-    height: 100vh;
-    width: 80%;
-    display: grid;
-    gap: 1em 1em;
-    padding-left: 1em;
-}
-  
-.history h2 {
-    font-family: 'Roboto', sans-serif;
-    font-size: 2rem;
-    padding-top: 1em;
-}
-  
-.history p {
-    font-family: 'Roboto', sans-serif;
-    font-size: 1.5rem;
-    margin-top: -9.5em;
-}
-  
-.history Image {
-    position: absolute;
-    top: 0em;
-    right: 3em;
-}
-
-/* Support */
-
-.about {
-    float: left;
-    height: 100vh;
-    width: 80%;
-    display: grid;
-    gap: 1em 1em;
-    padding-left: 1em;
-}
-  
-.about h2 {
-    font-family: 'Roboto', sans-serif;
-    font-size: 2rem;
-    padding-top: 1em;
-}
-  
-.about p {
-    font-family: 'Roboto', sans-serif;
-    font-size: 1.5rem;
-    /*margin-top: 3em;*/
-}
-  
-.about Image {
-    position: absolute;
-    top: 0em;
-    right: 3em;
-    object-fit: cover;
-}
-  
-.infoCards {
-    display: grid;
-    gap: 1em 1em;
-    padding-left: 1em;
-    padding-top: 2rem;
-    padding-bottom: 1rem;
-}
-  
-.card1, .card2, .card3, .card4 {
-    width: 80%;
-    height: 100%;
-}
-  
-.card1 {
-    grid-column-start: 1;
-    grid-column-end: 2;
-    grid-row-start: 1;
-    grid-row-end: 2;
-}
-  
-.card2 {
-    grid-column-start: 3;
-    grid-column-end: 4;
-    grid-row-start: 1;
-    grid-row-end: 2;
-}
-  
-.card3 {
-    grid-column-start: 1;
-    grid-column-end: 2;
-    grid-row-start: 2;
-    grid-row-end: 3;
-}
-  
-.card4 {
-    grid-column-start: 3;
-    grid-column-end: 4;
-    grid-row-start: 2;
-    grid-row-end: 3;
-}
-
-/* Shopping Lists */
-
-.header {
-    font-family: 'Roboto', sans-serif;
-    padding-bottom: 15px;
-    font-size: 50px;
-    text-align: center;
-}  
-
-.subheader {
-    font-family: 'Roboto', sans-serif;
-    padding-bottom: 15px;
-    font-size: 25px;
-    text-align: center;
-}
-
-.button {
-    font-family: 'Roboto', sans-serif;
-    padding: 3px 10px;
-    font-size: 12px;
-    border-radius: 3px;
-    border: 1px solid #e1e4e8;
-    text-align: center;
-  }
-
