+++ styles/Home.module.scss
@@ -261,6 +261,294 @@
   border: 0.1rem solid #B5B0AF;
 }
 
+/* Account & ActMenu */
+
+.no_padding_container {
+  padding: 0;
+  position: sticky;
+}
+
+.menu1 {
+  height: 100vh;
+  text-align: center;
+  width: 15%;
+  float: left;
+  background: linear-gradient(rgb(244, 250, 255), rgb(255, 255, 255));
+}
+
+.tabs {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 10%;
+}
+
+.profile {
+  float: left;
+  height: 100vh;
+  width: 80%;
+  display: grid;
+  gap: 1em 1em;
+  padding-left: 1em;
+}
+
+.grid1 {
+  grid-column-start: 1;
+  grid-column-end: 2;
+  grid-row-start: 1;
+  grid-row-end: 2;
+  padding-top: 3em;
+}
+
+.grid2 {
+  grid-column-start: 1;
+  grid-column-end: 2;
+  grid-row-start: 2;
+  grid-row-end: 3;
+}
+
+.grid3 {
+  grid-column-start: 2;
+  grid-column-end: 3;
+  grid-row-start: 2;
+  grid-row-end: 3;
+}
+
+.grid4 {
+  grid-column-start: 3;
+  grid-column-end: 4;
+  grid-row-start: 2;
+  grid-row-end: 3;
+}
+
+.grid5 {
+  grid-column-start: 1;
+  grid-column-end: 2;
+  grid-row-start: 3;
+  grid-row-end: 4;
+}
+
+.grid6 {
+  grid-column-start: 2;
+  grid-column-end: 3;
+  grid-row-start: 3;
+  grid-row-end: 4;
+}
+
+.grid7 {
+  grid-column-start: 3;
+  grid-column-end: 4;
+  grid-row-start: 3;
+  grid-row-end: 4;
+}
+
+.grid8 {
+  grid-column-start: 1;
+  grid-column-end: 2;
+  grid-row-start: 4;
+  grid-row-end: 5;
+  align-content: center;
+}
+
+.grid9 {
+  grid-column-start: 3;
+  grid-column-end: 4;
+  grid-row-start: 1;
+  grid-row-end: 2;
+  padding-top: 1em;
+}
+
+.kartLogo {
+  width: 150px;
+  height: 150px;
+}
+
+.grid1 h2 {
+  font-family: 'Roboto', sans-serif;
+  font-size: 2rem;
+}
+
+.grid2, .grid3, .grid4, .grid5, .grid6, .grid7, .grid8 label {
+  font-family: 'Roboto', sans-serif;
+}
+
+.updateBtn:hover {
+  background-color: white;
+  color: blue;
+}
+
+.menuBtn:hover {
+  background-color: linear-gradient(rgb(244, 250, 255), rgb(255, 255, 255));
+  color: white;
+}
+
+.menuBtn:active {
+  background-color: blue;
+  color: white;
+  box-shadow: 0 5px #666;
+  transform: translateY(4px);
+}
+
+/* Account History */
+
+.history {
+  float: left;
+  height: 100vh;
+  width: 80%;
+  display: grid;
+  gap: 1em 1em;
+  padding-left: 1em;
+}
+
+.history h2 {
+  font-family: 'Roboto', sans-serif;
+  font-size: 2rem;
+  padding-top: 1em;
+}
+
+.history p {
+  font-family: 'Roboto', sans-serif;
+  font-size: 1.5rem;
+  margin-top: -9.5em;
+}
+
+.history Image {
+  position: absolute;
+  top: 0em;
+  right: 3em;
+}
+
+/* Account About */
+
+.about {
+  float: left;
+  height: 100vh;
+  width: 80%;
+  display: grid;
+  gap: 1em 1em;
+  padding-left: 1em;
+}
+
+.about h2 {
+  font-family: 'Roboto', sans-serif;
+  font-size: 2rem;
+  padding-top: 1em;
+}
+
+.about p {
+  font-family: 'Roboto', sans-serif;
+  font-size: 1.5rem;
+  /*margin-top: 3em;*/
+}
+
+.about Image {
+  position: absolute;
+  top: 0em;
+  right: 3em;
+  object-fit: cover;
+}
+
+.infoCards {
+  display: grid;
+  gap: 1em 1em;
+  padding-left: 1em;
+  padding-top: 2rem;
+  padding-bottom: 1rem;
+}
+
+.card1, .card2, .card3, .card4 {
+  width: 80%;
+  height: 100%;
+}
+
+.card1 {
+  grid-column-start: 1;
+  grid-column-end: 2;
+  grid-row-start: 1;
+  grid-row-end: 2;
+}
+
+.card2 {
+  grid-column-start: 3;
+  grid-column-end: 4;
+  grid-row-start: 1;
+  grid-row-end: 2;
+}
+
+.card3 {
+  grid-column-start: 1;
+  grid-column-end: 2;
+  grid-row-start: 2;
+  grid-row-end: 3;
+}
+
+.card4 {
+  grid-column-start: 3;
+  grid-column-end: 4;
+  grid-row-start: 2;
+  grid-row-end: 3;
+}
+
+/* Account Privacy */
+
+.privacy {
+  float: left;
+  height: 100vh;
+  width: 80%;
+  display: grid;
+  gap: 1em 1em;
+  padding-left: 1em;
+}
+
+.privacy h2 {
+  font-family: 'Roboto', sans-serif;
+  font-size: 2rem;
+  padding-top: 1em;
+}
+
+.privacy p {
+  font-family: 'Roboto', sans-serif;
+  font-size: 1.5rem;
+  margin-top: 3em;
+  padding-bottom: 1em;
+}
+
+.privacy Image {
+  position: absolute;
+  top: 0em;
+  right: 3em;
+}
+
+/* Account Security */
+
+.security {
+  float: left;
+  height: 100vh;
+  width: 80%;
+  display: grid;
+  gap: 1em 1em;
+  padding-left: 1em;
+}
+
+.security h2 {
+  font-family: 'Roboto', sans-serif;
+  font-size: 2rem;
+  padding-top: 2em;
+}
+
+.security p {
+  font-family: 'Roboto', sans-serif;
+  font-size: 1.5rem;
+  margin-top: 1em;
+}
+
+.security Image {
+  position: absolute;
+  top: 0em;
+  right: 3em;
+}
+
 /* Checkout */
 
 .checkoutBox {
